+++ doc/changelog.md
@@ -5,7 +5,7 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* Fix: Exception when accessing context.Features